Remove all compilation warnings.  The only warnings left at build time are: -   if -jn n>1 is used then compiling the sbs driver will note that     only serial compilation works for that make. -   The swig compiler warns about not being able to support chunks of     the CXXUsb header files.

+++ main/base/os/daqshm.cpp
@@ -246,8 +246,10 @@ CDAQShm::attach(std::string name)
   close(fd);
   
   // Add a new mapping entry with refcount 1:
-
-  attachInformation initialInfo = {pMemory, fileSize, 1};
+  // Note that fileSize is only negative in errors which were already handled.
+  // so the cast below is safe as fileSize will be size_t  by now.
+  
+  attachInformation initialInfo = {pMemory, static_cast<size_t>(fileSize), 1};
   m_attachMap[name] = initialInfo;
 
   return pMemory;
